Change to slurp
If there are non-UTF-8 characters in perl comments, the following waring
will be shown during testing:

UTF-8 "\xE9" does not map to Unicode at .../lib/perl5/Path/Tiny.pm line 2077.

As an example, that message came with a file having the following lines
in it:

    # STDOUT on French machine:
    # [SC] EnumQueryServicesStatus:OpenService échec(s) 1060 :

The second line has "é" in it, which is "Latin Small Letter E With
Acute".

I was able to change to `path->($filename)->slurp` to resolve the issue.

Since it seems comments _could_ contain non-UTF-8, I think this should
do a simple slurp instead of slurp_utf8.